+++ src/main/java/com/github/nalamodikk/common/block/blockentity/mana_plate_press/ManaPlatePressScreen.java
@@ -9,7 +9,6 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
-
 import java.util.List;
 
 public class ManaPlatePressScreen extends AutoSizedModularScreen<ManaPlatePressMenu> {