+++ src/main/java/com/github/nalamodikk/common/inventory/sort/SortMode.java
@@ -1,0 +1,10 @@
+package com.github.nalamodikk.common.inventory.sort;
+
+public enum SortMode {
+    BY_TYPE, BY_NAME, BY_COUNT;
+
+    public SortMode next() {
+        SortMode[] values = values();
+        return values[(ordinal() + 1) % values.length];
+    }
+}