+++ src/main/java/com/github/nalamodikk/common/inventory/sort/SortTarget.java
@@ -1,0 +1,5 @@
+package com.github.nalamodikk.common.inventory.sort;
+
+public enum SortTarget {
+    CONTAINER, PLAYER_INVENTORY
+}